refactor: userType, authority common에서 제거
