--- common/src/main/java/org/bogus/groove/common/UserInfo.java
@@ -1,16 +1,0 @@
-package org.bogus.groove.common;
-
-import java.util.List;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.bogus.groove.common.enumeration.Authority;
-import org.bogus.groove.common.enumeration.UserType;
-
-@Getter
-@RequiredArgsConstructor
-public class UserInfo {
-    private final Long id;
-    private final String email;
-    private final UserType type;
-    private final List<Authority> authorities;
-}
