--- common/src/main/java/org/bogus/groove/common/enumeration/Authority.java
@@ -1,11 +1,0 @@
-package org.bogus.groove.common.enumeration;
-
-import java.util.List;
-
-public enum Authority {
-    USER,
-    TRAINER,
-    ADMIN;
-
-    public static final List<Authority> DEFAULT = List.of(USER);
-}
