--- groove-api/src/main/java/org/bogus/groove/domain/user/UserType.java
@@ -1,10 +1,0 @@
-package org.bogus.groove.domain.user;
-
-public enum UserType {
-    GROOVE,
-    KAKAO,
-    NAVER,
-    GOOGLE;
-
-    public static final UserType DEFAULT = GROOVE;
-}
