+++ groove-api/src/main/java/org/bogus/groove/endpoint/middleware/Authorized.java
@@ -4,7 +4,7 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import org.bogus.groove.common.enumeration.Authority;
+import org.bogus.groove.client.auth.Authority;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
