+++ groove-auth-api/src/main/java/org/bogus/groove_auth/endpoint/user/UserController.java
@@ -2,8 +2,9 @@ package org.bogus.groove_auth.endpoint.user;
 
 import lombok.RequiredArgsConstructor;
 import org.bogus.groove.common.CommonResponse;
-import org.bogus.groove.common.UserInfo;
+import org.bogus.groove_auth.domain.user.UserInfo;
 import org.bogus.groove_auth.domain.user.UserService;
+import org.bogus.groove_auth.domain.user.authority.Authority;
 import org.bogus.groove_auth.endpoint.auth.RegisterRequest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.GetMapping;
