+++ groove-auth/src/main/java/org/bogus/groove_auth/domain/user/authority/UserAuthorityUpdater.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.bogus.groove.common.enumeration.Authority;
 import org.bogus.groove_auth.storage.UserAuthorityEntity;
 import org.bogus.groove_auth.storage.UserAuthorityRepository;
 import org.springframework.stereotype.Component;
